chore: split db file into multiple files and centralized pagination
